workerManager no multi-task
